feat : Commands and Queries of Bookings and Homes implemented and completed using Redis and repository implemented migration done

+++ src/01.Core/HouseRent.Core.Domain/Amenities/Events/AmenityCreated.cs
@@ -4,7 +4,7 @@ using HouseRent.Core.Domain.Shared.ValueObjects;
 
 namespace HouseRent.Core.Domain.Amenities.Events;
 
-public record AmenityCreated(int id,
-                          Title title,
-                          Description description,
-                          Money money) : IDomainEvent;
+public record AmenityCreated(long id,
+                             Title title,
+                             Description description,
+                             Money money) : IDomainEvent;
