+++ src/01.Core/HouseRent.Core.Domain/Bookings/Events/BookingConfirmedDomainEvent.cs
@@ -1,4 +1,4 @@
 ﻿using HouseRent.Core.Domain.Framework;
 
 namespace HouseRent.Core.Domain.Bookings.Events;
-public sealed record BookingConfirmedDomainEvent(int BookingId) : IDomainEvent;
+public sealed record BookingConfirmedDomainEvent(long BookingId) : IDomainEvent;
