+++ src/01.Core/HouseRent.Core.Domain/Bookings/Parameters/CreateBookingParameters.cs
@@ -3,9 +3,9 @@ using HouseRent.Core.Domain.Bookings.ValueObjects;
 using HouseRent.Core.Domain.Shared.ValueObjects;
 
 namespace HouseRent.Core.Domain.Bookings.Parameters;
-public record CreateBookingParameters(int Id,
-                                     int HomeId,
-                                     int UserId,
+public record CreateBookingParameters(long Id,
+                                     long HomeId,
+                                     long UserId,
                                      DateRange Duration,
                                      Money PriceForPeriod,
                                      Money AmenitiesUpCharge,
